+++ plataformer/timer.cpp
@@ -1,8 +1,8 @@
 #include "include/timer.h"
 #include "raylib.h"
-void StartTimer(Timer* timer, double lifetime)
+void StartTimer(Timer* timer, float lifetime)
 {
-    timer->StartTime = GetTime();
+    timer->StartTime = float(GetTime());
     timer->Lifetime = lifetime;
 }
 
@@ -11,7 +11,7 @@ bool TimerDone(Timer* timer)
     return GetTime() - timer->StartTime >= timer->Lifetime;
 }
 
-double GetElapsed(Timer* timer)
+float GetElapsed(Timer* timer)
 {
     return GetTime() - timer->StartTime;
 }
